face_recognition/requirements-conda.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.4
cachetools==5.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cmake==3.22.4
colorama==0.4.4
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.4
django-crispy-forms==1.14.0
django-filter==21.1
djangorestframework==3.13.1
dlib==19.24.0
entrypoints==0.4
executing==0.8.3
face-recognition==1.4.0
face-recognition-models==0.3.0
fastjsonschema==2.15.3
filelock==3.7.0
Flask==2.1.2
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.11.3
imutils==0.5.4
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
MouseInfo==0.1.3
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
pure-eval==0.2.2
pyarrow==8.0.0
PyAutoGUI==0.9.53
pycparser==2.21
pydeck==0.7.1
PyGetWindow==0.0.9
Pygments==2.12.0
Pympler==1.0.1
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.3
pywin32==304
pywinpty==2.0.5
pyzmq==22.3.0
requests==2.27.1
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
streamlit==1.9.0
streamlit-lottie==0.0.3
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==5.2.1.post0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.19.0
virtualenv==20.14.1
watchdog==2.1.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.0
wikipedia==1.4.0
zipp==3.8.0
